Move game states into their own .c files
