+++ source/editor/main.c
@@ -5,6 +5,13 @@
 #include "../level.h"
 #include "../windows/debug_layer.h"
 #include "../windows/psx.h"
+#include "../main.h"
+
+int widescreen = 0;
+state_t current_state = STATE_NONE;
+state_t prev_state = STATE_NONE;
+int should_transition_state = 0;
+state_vars_t state;
 
 int main(void) {
     renderer_init();
